working properly (!) added CheckboxVertical for languages
