+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=${PORT:8080}
+server.port=${PORT:8181}
 logging.level.org.atmosphere = warn
 spring.mustache.check-template-location = false
